:white_check_mark:Complete:Create functiom SafeDiv

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -7,9 +7,9 @@ CREATE FUNCTION SafeDiv(
     IN p_a INT,
     IN p_b INT
 )
-RETURNS FLOAT
+RETURNS INT
 BEGIN
-    DECLARE result FLOAT;
+    DECLARE result INT;
 
     -- Compute the result
     IF p_b == 0 THEN
